refactor(fixpnt): split fixpnt into core / manipulators / iostream layers (#1334)
Phase 2 continues after takum.

  header                              lines  headers   I/O
  fixpnt/fixpnt.hpp   (before)       108858      474     5
  fixpnt/core.hpp     (new)           78123      358     0

78123 is the blocktriple plateau, as expected -- fixpnt uses it, so it lands with cfloat
(78330), posit (77430) and lns (74762) rather than with takum (49429) and integer
(48328). That split is now firmly a blocktriple boundary, not a per-type property.

Preceded by the #1389 sweep with the hardened script: 32 gaps across 26 headers.

Out of the core: convert_to_decimal_string, operator<<, parse, operator>>, to_binary,
to_native and to_triple -- one contiguous 271-line block -- to a new iostream.hpp.
parse() goes with the stream operators rather than manipulators because it is their only
caller and needs <regex> and <sstream>. support/decimal.hpp follows
convert_to_decimal_string. native/ieee754.hpp becomes ieee754_core.hpp. Four std::cerr/
std::cout diagnostics become fprintf(stderr, ...).

THREE PRE-EXISTING DEFECTS FOUND, all of which the split exposed and none of which is
fixpnt-specific:

  - fixpnt/numeric_limits.hpp included fixpnt.hpp, its own UMBRELLA. That is circular,
    and it meant core.hpp -> numeric_limits.hpp -> fixpnt.hpp -> iostream.hpp: the core
    dragged in the entire text layer it had just been separated from. Now includes
    fixpnt_impl.hpp. areal/numeric_limits.hpp has the same circularity and is left for
    the areal split.
  - fixpnt/table.hpp had NO #pragma once. It is the only table.hpp in the tree without
    one; nine other headers elsewhere are also missing it.
  - utility/color_print.hpp had NO INCLUDES AT ALL while using std::ostream twice and
    std::setw once, relying entirely on whatever preceded it. Every number system's
    manipulators.hpp includes it, so it worked only because the umbrellas pulled the
    stream headers first. Now includes <ostream> and <iomanip>.

Also removed native/integers.hpp from fixpnt_impl.hpp, unused by the core, and
temporarily removed <map> with it -- std::map is used by the core's charLookup in
assign(), and the compiler caught it immediately. Restored.

Verified: both gates pass on gcc and clang; fixpnt suite 41 pass 0 fail; every api.cpp
55 pass 0 fail -- the relevant number here, since the color_print.hpp change touches
every number system; the hardened sweep reports 0 gaps across all 28 fixpnt headers.
