+++ include/sw/universal/number/fixpnt/attributes.hpp
@@ -5,6 +5,10 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <string>        // std::string
+#include <sstream>       // std::stringstream
+#include <iomanip>       // std::setw
+#include <cassert>       // assert()
 #include <cmath> // for std:pow()
 
 namespace sw { namespace universal {
