+++ include/sw/universal/number/fixpnt/core.hpp
@@ -1,0 +1,26 @@
+#pragma once
+// core.hpp: the fixpnt arithmetic core -- no I/O, no text
+//
+// Copyright (C) 2017 Stillwater Supercomputing, Inc.
+// SPDX-License-Identifier: MIT
+//
+// This file is part of the universal numbers project, which is released under an MIT Open Source license.
+//
+// Layer 1 of the fixpnt headers (#1334, Phase 2).
+//
+//     #include <universal/number/fixpnt/fixpnt.hpp>   // everything, as before
+//     #include <universal/number/fixpnt/core.hpp>     // arithmetic only
+//
+// NOTE: traits/arithmetic_traits.hpp and common/number_traits_reports.hpp are
+// deliberately NOT included -- they build report strings and pull <sstream>/<iomanip>.
+#include <universal/utility/architecture.hpp>
+#include <universal/utility/bit_cast.hpp>
+#include <universal/utility/long_double.hpp>
+
+#include <universal/traits/number_traits.hpp>
+
+#include <universal/number/fixpnt/exceptions.hpp>
+#include <universal/number/fixpnt/fixpnt_fwd.hpp>
+#include <universal/number/fixpnt/fixpnt_impl.hpp>
+#include <universal/traits/fixpnt_traits.hpp>
+#include <universal/number/fixpnt/numeric_limits.hpp>
