+++ include/sw/universal/number/fixpnt/fixpnt_fwd.hpp
@@ -6,6 +6,7 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 
+#include <string>        // std::string
 namespace sw { namespace universal {
 
 	// forward references
