+++ include/sw/universal/number/fixpnt/math/trigonometry.hpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <cmath>         // the <cmath> functions used below
 #include <math/constants/double_constants.hpp>
 
 namespace sw { namespace universal {
