+++ include/sw/universal/number/fixpnt/table.hpp
@@ -1,8 +1,13 @@
+#pragma once
 // table.hpp: generate a table of encoding and values for fixed-size arbitrary fixed-point configurations
 //
 // Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+
+#include <cstddef>     // std::size_t
+#include <ostream>     // std::ostream
+#include <iomanip>     // std::setw
 
 namespace sw { namespace universal {
